+++ CHANGELOG.rst
@@ -8,6 +8,18 @@ Changelog
 =========
 
 
+4.2.0 - 2026-01-13
+------------------
+
+**New feature:**
+
+- :func:`tabmat.from_formula` now also supports any dataframe supported by narwhals.
+
+**Other changes:**
+
+- Require Python>=3.10.
+
+
 4.1.5 - 2025-12-17
 ------------------